mixup running

+++ src/torch_datasets/miniworld/base_city.py
@@ -9,11 +9,11 @@ class BaseCity(Base, ABC):
     @staticmethod
     def colors_to_labels(labels_color):
 
-        labels0 = np.zeros(shape=labels_color.shape[:2], dtype=int)
-        labels1 = np.zeros(shape=labels_color.shape[:2], dtype=int)
+        labels0 = np.zeros(shape=labels_color.shape[:2], dtype=float)
+        labels1 = np.zeros(shape=labels_color.shape[:2], dtype=float)
         mask = np.any(labels_color != [0], axis=2)
-        np.putmask(labels0, ~mask, 1)
-        np.putmask(labels1, mask, 1)
+        np.putmask(labels0, ~mask, 1.)
+        np.putmask(labels1, mask, 1.)
         labels = np.stack([labels0, labels1], axis=2)
 
         return labels
